Moved all functions out of main
