+++ Core/User/FSK.c
@@ -9,7 +9,10 @@
 #define dty_cyc1 (freq_1/2)
 #define dty_cyc0 (freq_0/2)
 
-void init_fsk(uint8_t* msg, uint32_t* pwm_data){
+uint32_t pwm_data[24];
+int data_pos = 0;
+
+void init_fsk(uint8_t* msg){
     int count = 0;
     for(int j=0; j < 3; j++){
         for(int i=0; i < 8; i++){
@@ -24,7 +27,7 @@ void init_fsk(uint8_t* msg, uint32_t* pwm_data){
     }
 }
 
-void start_fsk(uint32_t* pwm_data){
+void start_fsk(void){
     __HAL_TIM_SET_AUTORELOAD(&htim2, pwm_data[0]);
     __HAL_TIM_SET_COUNTER(&htim2, 0);
     __HAL_TIM_SET_COUNTER(&htim15, 0);
@@ -35,4 +38,18 @@ void start_fsk(uint32_t* pwm_data){
 void stop_fsk(void){
     HAL_TIM_Base_Stop_IT(&htim15);
     HAL_TIM_PWM_Stop(&htim2,TIM_CHANNEL_2);
+}
+
+void HAL_TIM_PeriodElapsedCallback(TIM_HandleTypeDef *htim){
+    if(htim->Instance == TIM15){
+        data_pos++;
+        if(data_pos > 23){
+            stop_fsk();
+            data_pos = 0;
+        }else{
+            __HAL_TIM_SET_AUTORELOAD(&htim2, pwm_data[data_pos]);
+            __HAL_TIM_SET_COUNTER(&htim2, 0);
+            __HAL_TIM_SET_COMPARE(&htim2, TIM_CHANNEL_2, pwm_data[data_pos]/2);
+        }
+    }
 }
